refactor(models): move schema, parameters and paginated response to models_utils
